refactor(modeling) renamed measureBounds to measureBoundingBox
