+++ packages/modeling/src/measurements/measureBoundingBox.js
